--- src/main/java/xyz/lannt/task/application/property/DataCollectorProperty.java
@@ -1,16 +1,0 @@
-package xyz.lannt.task.application.property;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-import lombok.Data;
-
-@Data
-@Component
-@ConfigurationProperties(prefix = "data-collector")
-public class DataCollectorProperty {
-
-  private int size;
-
-  private int delay;
-}
